tech-docs-qa/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
langchain==0.0.XXX  # replace with the version you prefer
openai==0.29.0
qdrant-client==1.3.1
sentence-transformers==2.2.2
transformers==4.35.2
python-dotenv==1.0.0
pydantic==1.10.7
requests==2.31.0
pytest==7.4.0
scikit-learn==1.3.0
spacy>=3.5.0; extra == "nlp"